fix: make contract compile without middleware dep
move over the test using middleware to ethers-middleware
